Funcion al boton Eliminar registros

+++ scrips/DeleteUsuario.php
@@ -1,13 +1,17 @@
 <?php
     require_once("conexion.php");
-
-    $idUsuario = $_GET['id'];    
+    if(isset($_POST['DeleteCambios']))
+    {
+        $idUsuario = $_POST['idUsuarioD'];    
     
-    $sql = "DELETE FROM usuarios WHERE idUsuario = $idUsuario";
-    $resultado = $mysqli->query($sql);
+        $sql = "DELETE FROM usuarios WHERE idUsuario = $idUsuario";
+        $resultado = $mysqli->query($sql);
 
-    if($resultado)
-    { header("Location: ../MenuPrincipal/MainUsuarios.php"); }
-    else { echo "Error al eliminar."; }
+        if($resultado)
+        { header("Location: ../MenuPrincipal/MainUsuarios.php"); }
+        else { echo "Error al eliminar."; }
+    }
+
+    
     
 ?>
